rename folders to lower case
